+++ ak_library_management/__manifest__.py
@@ -2,7 +2,7 @@
 {
     'name': 'Library Management',
     'author': 'Moin Musa',
-    'version': '18.0.1.1.8',
+    'version': '18.0.1.1.9',
     'summary': 'Manages books,books category and members at library',
     'website': 'https://www.aktivsoftware.com',
     'description': """
@@ -10,7 +10,7 @@
     Manges books with its categories.
     Manges member at library.
     """,
-    'depends': ['stock', 'contacts', 'sale_management', 'base_automation', 'hr', 'website', 'point_of_sale'],
+    'depends': ['stock', 'contacts', 'sale_management', 'base_automation', 'hr', 'website_sale', 'point_of_sale'],
     'data': [
         'security/security.xml',
         'security/ir.model.access.csv',
@@ -39,6 +39,7 @@
         'report/report_custom_invoice.xml',
         'views/contacts_template.xml',
         'views/customer_page.xml',
+        'views/website_templates.xml',
     ],
     'assets':{
         'web.assets_frontend': [