+++ ak_library_management/changelog.md
@@ -57,3 +57,7 @@
 
 ## [18.0.1.1.8] - 2025-04-01 | Front-end - Practical Task (POS 18 products section)
 - add and inherit product card and product screen template.
+
+## [18.0.1.1.9] - 2025-04-02 | Front-end - Practical Task (Download Product Images)
+- add controller for fetch images of products from the product template
+- extend view of website product template